Modificaciones de la librería para compatibilidad con linux. La librería ya compila en linux pero no se ha probado aun.

+++ src/fourier.cpp
@@ -162,7 +162,7 @@ double fourierLinesDetection(cv::Mat &source, std::vector<int> &cols, std::vecto
 
   pts->clear();
   pts->resize(cols.size());
-  for (int j = 0; j < cols.size(); j++) {
+  for (size_t j = 0; j < cols.size(); j++) {
 
     for (int i = 0; i < size.height; i++) {
       curVar = imageOut.at<float>(cv::Point(cols[j], i)) - imageOut.at<float>(cv::Point(cols[j], 0));
